+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Updated SDK constraint, lints and format.
 - `BitArray.setWhere` and `BitArray.clearWhere` for callback-driven updates.
+- `BitArray.setRange` and `BitArray.clearRange` for efficient continous-range updates.
 - More efficient `BitArray` operations: `setAll`, `clearAll`.
 - More efficient `BitArray` operations: `and`, `andNot`, `or`, `xor` when the parameter is another `BitArray`.